+++ lib/AST/MemoryAccess.cpp
@@ -593,7 +593,7 @@ Expr *ASTTranslate::accessMemTexAt(DeclRefExpr *LHS, HipaccAccessor *Acc,
     args.push_back(LHStex);
     // byte addressing required for surf2Dwrite
     args.push_back(createBinaryOperator(Ctx, idx_x, createIntegerLiteral(Ctx,
-            (int)Acc->getImage()->getPixelSize()), BO_Mul, Ctx.IntTy));
+            (int32_t)Acc->getImage()->getPixelSize()), BO_Mul, Ctx.IntTy));
     args.push_back(idx_y);
   }
 
@@ -807,14 +807,14 @@ void ASTTranslate::stageIterationToSharedMemory(SmallVector<Stmt *, 16>
         if (compilerOptions.exploreConfig()) {
           SX2 = tileVars.local_size_x;
         } else {
-          SX2 = createIntegerLiteral(Ctx, (int)Kernel->getNumThreadsX());
+          SX2 = createIntegerLiteral(Ctx, (int32_t)Kernel->getNumThreadsX());
         }
       } else {
         SX2 = createIntegerLiteral(Ctx, 0);
       }
       if (Acc->getSizeY() > 1) {
         global_offset_y = createParenExpr(Ctx, createUnaryOperator(Ctx,
-              createIntegerLiteral(Ctx, (int)Acc->getSizeY()/2), UO_Minus,
+              createIntegerLiteral(Ctx, (int32_t)Acc->getSizeY()/2), UO_Minus,
               Ctx.IntTy));
       } else {
         global_offset_y = NULL;
@@ -834,7 +834,7 @@ void ASTTranslate::stageIterationToSharedMemory(SmallVector<Stmt *, 16>
         Expr *local_offset_x = NULL;
         if (Acc->getSizeX() > 1) {
           local_offset_x = createBinaryOperator(Ctx, createIntegerLiteral(Ctx,
-                i), tileVars.local_size_x, BO_Mul, Ctx.IntTy);
+                (int32_t)i), tileVars.local_size_x, BO_Mul, Ctx.IntTy);
           global_offset_x = createBinaryOperator(Ctx, local_offset_x, SX2,
               BO_Sub, Ctx.IntTy);
         }
@@ -870,7 +870,7 @@ void ASTTranslate::stageIterationToSharedMemoryExploration(SmallVector<Stmt *,
         if (compilerOptions.exploreConfig()) {
           SX2 = tileVars.local_size_x;
         } else {
-          SX2 = createIntegerLiteral(Ctx, (int)Kernel->getNumThreadsX());
+          SX2 = createIntegerLiteral(Ctx, (int32_t)Kernel->getNumThreadsX());
         }
       } else {
         SX2 = createIntegerLiteral(Ctx, 0);
@@ -880,7 +880,7 @@ void ASTTranslate::stageIterationToSharedMemoryExploration(SmallVector<Stmt *,
       if (Acc->getSizeY() > 1) {
         global_offset_y = createBinaryOperator(Ctx, global_offset_y,
             createUnaryOperator(Ctx, createIntegerLiteral(Ctx,
-                (int)Acc->getSizeY()/2), UO_Minus, Ctx.IntTy), BO_Add,
+                (int32_t)Acc->getSizeY()/2), UO_Minus, Ctx.IntTy), BO_Add,
             Ctx.IntTy);
       }
 
@@ -898,7 +898,7 @@ void ASTTranslate::stageIterationToSharedMemoryExploration(SmallVector<Stmt *,
         Expr *local_offset_x = NULL;
         if (Acc->getSizeX() > 1) {
           local_offset_x = createBinaryOperator(Ctx, createIntegerLiteral(Ctx,
-                i), tileVars.local_size_x, BO_Mul, Ctx.IntTy);
+                (int32_t)i), tileVars.local_size_x, BO_Mul, Ctx.IntTy);
           global_offset_x = createBinaryOperator(Ctx, local_offset_x, SX2,
               BO_Sub, Ctx.IntTy);
         }
@@ -914,7 +914,7 @@ void ASTTranslate::stageIterationToSharedMemoryExploration(SmallVector<Stmt *,
 
       Expr *SY;
       if (Kernel->getPixelsPerThread() > 1) {
-        SY = createIntegerLiteral(Ctx, (int)Kernel->getPixelsPerThread());
+        SY = createIntegerLiteral(Ctx, (int32_t)Kernel->getPixelsPerThread());
       } else {
         SY = createIntegerLiteral(Ctx, 1);
       }
@@ -922,7 +922,7 @@ void ASTTranslate::stageIterationToSharedMemoryExploration(SmallVector<Stmt *,
       if (Acc->getSizeY() > 1) {
         SY = createBinaryOperator(Ctx, SY, createBinaryOperator(Ctx,
               createBinaryOperator(Ctx, createIntegerLiteral(Ctx,
-                  (int)Acc->getSizeY()-2), DSY, BO_Div, Ctx.IntTy),
+                  (int32_t)Acc->getSizeY()-2), DSY, BO_Div, Ctx.IntTy),
               createIntegerLiteral(Ctx, 1), BO_Add, Ctx.IntTy), BO_Add,
             Ctx.IntTy);
       }
